+++ src/crypto/mod.rs
@@ -1,0 +1,2 @@
+/// provides support for calculating digests (`sha256`/`sha512`) - requires `crypto` feature
+pub mod digest;