+++ src/lib.rs
@@ -1,7 +1,9 @@
-#[cfg(test)]
-mod tests {
-    #[test]
-    fn it_works() {
-        assert_eq!(2 + 2, 4);
-    }
-}
+/// a collection of extensions for filesystem operations
+pub mod extensions;
+
+/// a collection of generic types/bounds
+pub mod generics;
+
+#[cfg(feature = "crypto")]
+/// support for cryptographic operations (requires `crypto` features)
+pub mod crypto;